:art: 헥사고날 아키텍처 적용
