+++ src/main/java/com/reachrich/reachrichuser/domain/exception/CustomException.java
@@ -1,0 +1,11 @@
+package com.reachrich.reachrichuser.domain.exception;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+@Getter
+public class CustomException extends RuntimeException {
+
+    private final ErrorCode errorCode;
+}
