+++ src/main/java/com/reachrich/reachrichuser/domain/user/LoginDto.java
@@ -1,0 +1,10 @@
+package com.reachrich.reachrichuser.domain.user;
+
+import lombok.Getter;
+
+@Getter
+public class LoginDto {
+
+    private String email;
+    private String password;
+}
