+++ src/main/java/com/reachrich/reachrichuser/domain/user/LogoutDto.java
@@ -1,0 +1,9 @@
+package com.reachrich.reachrichuser.domain.user;
+
+import lombok.Getter;
+
+@Getter
+public class LogoutDto {
+
+    private String nickname;
+}
