+++ src/main/java/com/reachrich/reachrichuser/domain/user/RegisterDto.java
@@ -1,0 +1,12 @@
+package com.reachrich.reachrichuser.domain.user;
+
+import lombok.Getter;
+
+@Getter
+public class RegisterDto {
+
+    private String email;
+    private String password;
+    private String nickname;
+    private String authCode;
+}
