--- src/main/java/com/reachrich/reachrichuser/global/exception/CustomException.java
@@ -1,11 +1,0 @@
-package com.reachrich.reachrichuser.global.exception;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-@Getter
-public class CustomException extends RuntimeException {
-
-    private final ErrorCode errorCode;
-}
