--- src/main/java/com/reachrich/reachrichuser/global/exception/ErrorResponse.java
@@ -1,11 +1,0 @@
-package com.reachrich.reachrichuser.global.exception;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-@Getter
-public class ErrorResponse {
-
-    private final String message;
-}
