+++ src/main/java/com/reachrich/reachrichuser/infrastructure/repository/UserRepository.java
@@ -1,0 +1,14 @@
+package com.reachrich.reachrichuser.infrastructure.repository;
+
+import com.reachrich.reachrichuser.domain.user.User;
+import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+
+    Optional<User> findByEmail(String email);
+
+    boolean existsByEmail(String email);
+}
