--- src/main/java/com/reachrich/reachrichuser/user/dto/LoginDto.java
@@ -1,10 +1,0 @@
-package com.reachrich.reachrichuser.user.dto;
-
-import lombok.Getter;
-
-@Getter
-public class LoginDto {
-
-    private String email;
-    private String password;
-}
