--- src/main/java/com/reachrich/reachrichuser/user/dto/LogoutDto.java
@@ -1,9 +1,0 @@
-package com.reachrich.reachrichuser.user.dto;
-
-import lombok.Getter;
-
-@Getter
-public class LogoutDto {
-
-    private String nickname;
-}
