--- src/main/java/com/reachrich/reachrichuser/user/dto/RegisterDto.java
@@ -1,12 +1,0 @@
-package com.reachrich.reachrichuser.user.dto;
-
-import lombok.Getter;
-
-@Getter
-public class RegisterDto {
-
-    private String email;
-    private String password;
-    private String nickname;
-    private String authCode;
-}
