--- src/main/java/com/reachrich/reachrichuser/user/repository/UserRepository.java
@@ -1,14 +1,0 @@
-package com.reachrich.reachrichuser.user.repository;
-
-import com.reachrich.reachrichuser.user.entity.UserEntity;
-import java.util.Optional;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRepository extends JpaRepository<UserEntity, Long> {
-
-    Optional<UserEntity> findByEmail(String email);
-
-    boolean existsByEmail(String email);
-}
